requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
charset-normalizer==2.0.12
Django==4.0.3
idna==3.3
screen==1.0.1
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
xlwt==1.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt
django-cors-headers
django-ckeditor==6.2.0
django-allauth
django-extensions
pytils
redis